:checkered_flag: Start the Deck class, View inherits
