+++ TerminalPlayingCards/config.py
@@ -3,6 +3,7 @@
 # pylint: disable=missing-docstring
 from colorama import Fore, Back
 
+# Objects for the card module
 SUIT_SYMBOL_DICT = {
     "diamonds": {"symbol": "♦", "style": Back.WHITE + Fore.RED},
     "hearts": {"symbol": "♥", "style": Back.WHITE + Fore.RED},
@@ -54,3 +55,19 @@ CARD_FACE_DICT = {
 }
 
 CARD_BACK_STYLE = Fore.WHITE + Back.LIGHTBLACK_EX
+# Objects for the deck module
+DEFAULT_DECK_SPEC = {
+    "A": {"suit": {"clubs": 1, "diamonds": 1, "spades": 1, "hearts": 1}},
+    "2": {"suit": {"clubs": 2, "diamonds": 2, "spades": 2, "hearts": 2}},
+    "3": {"suit": {"clubs": 3, "diamonds": 3, "spades": 3, "hearts": 3}},
+    "4": {"suit": {"clubs": 4, "diamonds": 4, "spades": 4, "hearts": 4}},
+    "5": {"suit": {"clubs": 5, "diamonds": 5, "spades": 5, "hearts": 5}},
+    "6": {"suit": {"clubs": 6, "diamonds": 6, "spades": 6, "hearts": 6}},
+    "7": {"suit": {"clubs": 7, "diamonds": 7, "spades": 7, "hearts": 7}},
+    "8": {"suit": {"clubs": 8, "diamonds": 8, "spades": 8, "hearts": 8}},
+    "9": {"suit": {"clubs": 9, "diamonds": 9, "spades": 9, "hearts": 9}},
+    "10": {"suit": {"clubs": 10, "diamonds": 10, "spades": 10, "hearts": 10}},
+    "J": {"suit": {"clubs": 11, "diamonds": 11, "spades": 11, "hearts": 11}},
+    "Q": {"suit": {"clubs": 12, "diamonds": 12, "spades": 12, "hearts": 12}},
+    "K": {"suit": {"clubs": 13, "diamonds": 13, "spades": 13, "hearts": 13}},
+}
